% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.25.04.09.R
\name{plot.shared.GOs}
\alias{plot.shared.GOs}
\title{Plot shared GOs}
\usage{
\method{plot}{shared.GOs}(
  GOs.pick,
  cols = c("cyan4", "firebrick3"),
  labels = c("P. aeruginosa", "E. coli")
)
}
\arguments{
\item{x}{An object of class 'shared.GOs'.}

\item{...}{Additional arguments passed to plotting functions.}
}
\description{
Custom plot method for shared.GOs objects.
}
